Moved tests and examples to separate dirs
This significantly cleans up the root dir, and makes it easier to find
all examples or all test-related files.

The example and test files have a dependency on gdpc itself, which is
probably why they resided in the root directory before. This is now
handled in different way: there is now a requirements-dev.txt, which
includes the requirement `-e .`. This will install gdpc itself in
editable mode, so that it is available to the example and test files.

This way, end users with gdpc installed can still use the example files
as-is, and gdpc developers can still run them against the gdpc files in
their working directory.