Renamed Interface class to Editor
Too many things were called Interface:
- direct_interface.py
- interface.py
- interface.Interface

Conceptually, the "interface" should encompass all interaction with
Minecraft. The class that was previously called "Interface" was not
doing this: it only reflected *part* of the interface. Some
functionality, such as getting the buildarea, was not included in it.

"Editor" seems like an appropriate name, since it describes most
generally what the class is responsible for: editing the Minecraft
world. It might still be changed, though.

The Editor class could also be flattened down into a single Interface
class that includes all functionality. However, the Editor class already
has a lot of responsibilities. It cannot be broken down further because
its functions need to share some common data, such as the block cache,
but its responsibilities can at least be limited to editing.