Consistified interface response propagation
Previously, many functions attempted to propagate the return values and
error messages from the HTTP interface, but it was very inconsistent,
and deduced return types were all over the place.

Now, only direct_interface.py functions return the full server
responses. Functions from interface.py and derivatives such as
geometry.py generally only return whether a function succeeded fully.

It would be nice if we could propagate all error messages and return
values all the way to the high-level functions, but this is simply not
reasonably possible. For example, when placing blocks in a buffered
fasion, return values are only available once the buffer is flushed.

This commit also changes buffer behavior slightly: blocks that failed to
place are now re-buffered. Buffered block placing functions can now
report failure if the buffer could not be sufficiently emptied to make
room for the new blocks.

Finally, this commit significantly cleans up and improves documentation
for direct_interface.py. The direct_interface.sendBlocks function is
removed because it is not a direct wrapper: its functionality has
been moved to interface.py.