Redesigned block placing functions
Most important changes:
1. Removed the ability to place blocks at a Transform (they can now only
   be placed at a vector).
2. Moved scale logic to placeBlock(), removed placeStringGlobal and made
   the geometry functions that used it for optimization purposes just use
   placeBlock().

**Reasons for change 1:**

Placing blocks at a Transform was an extremely niche feature. The
winning entry to GDMC 2022, with >9000 LOC, used it only once. It can
easily be replaced by a Block.transformed() call.

Removing the ability to call placeBlock() with a Transform opens op the
possibility of accepting a *Sequence* of positions. This is an
optimization opportunity: when taking a Sequence of positions,
placeBlock() only needs to transform the Block once, and it can
temporarily enable buffering. Accepting a Sequence of TransformLike's
would come with the problem that each Transform might lead to a
different transformed Block, making the first optimization impossible.
An alternative would be to accept both a single Transform and a list of
vectors, but this would be rather convoluted, and difficult to
understand for new users.

Furthermore, accepting only a position (and not a Transform) brings
placeBlock() in line with the geometry functions: these also only accept
positions.

Finally, removing the ability to call with a Transform made it possible
to merge placeStringGlobal and placeBlockGlobal, making the code a bit
easier to understand.

**Reasons for change 2:**

Thanks to change 1, placeStringGlobal could be merged into
placeBlockGlobal, but this did not in principle mean that the scale
logic had to be moved to placeBlock().

Some geometry functions previously used an optimization: they only
transformed the "key points" of the geometric region, and then called
placeStringGlobal directly. This avoided repated Block and position
transformations. To do this, it was necessary for there to be a function
that did global block placement, but still executed scaling logic.

There was however a major bug in this system: the geometric region
between transformed key points is not always equivalent to the
transformed geometric region between key points. The two differ when
scaling is involved: a scaled line between key points is thicker than a
line between scaled key points.

Since the system was bugged, this optimization feature was simply
removed. As a bonus, this made it possible to remove the "scale"
parameter from placeBlockGlobal, making that function a bit simpler and
more in line with placeBlock().

Avoiding repeated Block transformations can also be achieved by making
placeBlock() accept a sequence of positions as described above. Only the
avoidance of position transformations is now impossible.

An alternative would have been to just remove the scaling functionality
altogether (leaving only flipping). Without scaling, geometry functions
could safely perform the described optimization with placeBlockGlobal.
Scaling is also rarely useful. I decided to keep it; I believe the
feature may still have merit, and the geometry optimization was
admittedly premature.

---

Other changes:
- Renamed block.name to block.id.
- Moved block id palette selection to Block.chooseId().