Adjusted buffering/caching interaction
Previously, buffered blocks were cached for block retrievel only if
caching was enabled. This had a problem: it makes using only buffering
and not caching non-transparent: the behavior of placing and then
getting at the same position may change. All Editor performance options
should be transparent so long as nothing but the Editor is changing the
world.

Now, buffered blocks are always used during block retrieval, even if
caching is off.

Advantage:
 - Buffering becomes transparent as long as nothing but the Editor is
   changing the world.

Disadvantage:
 - Block retrieval may be inaccurate if something other than the Editor
   is changing the world while only buffering is enabled.

In my opinion, this is worth it.