Flattened Interface and improved error handling
The Interface class only contained an url, which is not really enough to
motivate it being a class.

The original reason for making Interface a class was to make it possible
to create an alternative (duck-typed) interface classes which don't use
the HTTP interface, and pass those to Editor. That would have required
changing HTTP-specific parameters such as `retries` and `timeout` to
class state. There were however several reasons why this design was
problematic. It made HTTP error handling and detection more complicated
and it forced users to construct an instance themselves if they wanted to
change the url or some other setting.

An abstract Interface-like class can be added later if it turns out to be
useful. The code should not be made more complicated for a theoretical
feature. On top of that, abstraction is already possible at the level of
the HTTP interface.

This commit also improves error handling by defining some custom
exceptions and adding editor.checkConnection().