Replaced block-id palettes with Block palettes
Having one Block instance represent a palette of multiple blocks was
kind of a code smell, and caused various problems.

A disadvantage of full Block palettes, and the reason why this
change only happened now, is that every block in a palette needs
to be transformed on placement. With block-id palettes, this could be
avoided. Clean code design is however probably more important than this
small theoretical performance difference.