Convert Iterables to ndarrays with np.fromiter
Using np.array() to convert Iterables to ndarrays fails in the case of
Sets: Numpy will create an array with one element, that being the set
itself.
To prevent this, we need to use np.fromiter to convert from iterables.

Fixes #90.

In principle, we also need to do this when converting from VecLikes.
This might hurt performance though, so we have left this unchanged for
now.