Simplified installation instructions
Although the detailed information was nice, most of it was not really
essential, and it made installation look more complicated than it is.

The bit about `requirements.txt` was useful, but also added extra
complication. it would be preferable to just have more accurate
dependency version bounds in setup.py. Users often report problems on
the Discord, so we should be able to eventually iron out any version
bound mistakes.