Moved to the src/ layout
It seems that this layout has some advantages over the previous, "flat"
one, such as ensuring that tests always run with the installed package
and not the files in the working directory. Some tools also work better
with this layout.

https://blog.ionelmc.ro/2014/05/25/python-packaging/#the-structure
https://hynek.me/articles/testing-packaging/
https://packaging.python.org/en/latest/discussions/src-layout-vs-flat-layout/
https://github.com/pypa/packaging.python.org/issues/320
https://stackoverflow.com/a/72792078

There are no open PR's at the time of writing, so this is a good moment
to switch.