Removed requirements.txt
A requirements file is not really necessary for a library. The
requirements and their bounds are already listed in setup.py; the only
thing which requirements.txt adds is pinned versions for each
dependency for gdpc developers to use, but it seems better to just
install the latest version of each dependency and update setup.py if we
find out some depdendency version is incompatible.

requirements-dev.txt is still useful for listing development-only
requirements.