Changed all GDMC links to the wikidot site
The official competition website has been unavailable for some time now,
with the wikidot website acting as its temporary(?) replacement. The
dead website links are probably very confusing to new participants, so
this commit changes them to the wikidot site for now.

The link on PyPI will only update if we make a new release, but the
readme changes are immediately visible.